--- src/main/java/com/tencent/wxcloudrun/WxCloudRunApplication.java
@@ -1,14 +1,0 @@
-package com.tencent.wxcloudrun;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan(basePackages = {"com.tencent.wxcloudrun.dao"})
-public class WxCloudRunApplication {  
-
-  public static void main(String[] args) {
-    SpringApplication.run(WxCloudRunApplication.class, args);
-  }
-}
