--- src/main/java/com/tencent/wxcloudrun/controller/AdsInfoController.java
@@ -1,32 +1,0 @@
-
-package com.tencent.wxcloudrun.controller;
-
-import com.tencent.wxcloudrun.dto.AdsPageParam;
-import com.tencent.wxcloudrun.dto.PageDTO;
-import com.tencent.wxcloudrun.dto.Result;
-import com.tencent.wxcloudrun.model.AdsInfoEntity;
-import com.tencent.wxcloudrun.service.AdsInfoService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * @author tangsh
- * @date 2022/10/27
- */
-@RestController
-@RequestMapping("/front")
-public class AdsInfoController {
-
-    @Autowired
-    private AdsInfoService adsInfoService;
-
-    @PostMapping("/v1/ads/page")
-    public Result<PageDTO<AdsInfoEntity>> page(@RequestBody @Validated AdsPageParam param) {
-        PageDTO<AdsInfoEntity> pageResult = adsInfoService.page(param);
-        return Result.Success(pageResult);
-    }
-}
