--- src/main/java/com/tencent/wxcloudrun/dao/AdsInfoMapper.java
@@ -1,6 +1,0 @@
-package com.tencent.wxcloudrun.dao;
-
-import com.tencent.wxcloudrun.model.AdsInfoEntity;
-
-public interface AdsInfoMapper extends DefaultMapper<AdsInfoEntity>{
-}
