--- src/main/java/com/tencent/wxcloudrun/dao/DefaultMapper.java
@@ -1,6 +1,0 @@
-package com.tencent.wxcloudrun.dao;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-public interface DefaultMapper<T> extends BaseMapper<T> {
-}
