--- src/main/java/com/tencent/wxcloudrun/model/BaseDO.java
@@ -1,21 +1,0 @@
-
-package com.tencent.wxcloudrun.model;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableLogic;
-import lombok.Data;
-
-import java.io.Serializable;
-import java.util.Date;
-
-@Data
-public abstract class BaseDO implements Serializable {
-
-    @TableId(type = IdType.AUTO)
-    private Integer id;
-    private Date created;
-    private Date modified;
-    @TableLogic
-    private Integer flag;
-}
