--- src/main/java/com/tencent/wxcloudrun/model/Counter.java
@@ -1,18 +1,0 @@
-package com.tencent.wxcloudrun.model;
-
-import lombok.Data;
-
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
-@Data
-public class Counter implements Serializable {
-
-  private Integer id;
-
-  private Integer count;
-
-  private LocalDateTime createdAt;
-
-  private LocalDateTime updatedAt;
-}
