--- src/main/java/com/tencent/wxcloudrun/repository/AdsInfoRepository.java
@@ -1,10 +1,0 @@
-package com.tencent.wxcloudrun.repository;
-
-import com.tencent.wxcloudrun.dao.AdsInfoMapper;
-import com.tencent.wxcloudrun.model.AdsInfoEntity;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public class AdsInfoRepository extends BaseRepository<AdsInfoMapper, AdsInfoEntity> {
-}
