--- src/main/java/com/tencent/wxcloudrun/service/CounterService.java
@@ -1,15 +1,0 @@
-package com.tencent.wxcloudrun.service;
-
-import com.tencent.wxcloudrun.model.Counter;
-
-import java.util.Optional;
-import java.util.List;
-
-public interface CounterService {
-
-  Optional<Counter> getCounter(Integer id);
-
-  void upsertCount(Counter counter);
-
-  void clearCount(Integer id);
-}
