+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/entity/AdsInfoEntity.java
@@ -1,0 +1,15 @@
+package com.tencent.wxcloudrun.entity;
+
+import com.baomidou.mybatisplus.annotation.TableName;
+import lombok.Data;
+
+@Data
+@TableName("ads_info")
+public class AdsInfoEntity extends BaseDO {
+
+    private String title;
+    private String des;
+    private String img;
+    private String content;
+    private String status;
+}
