+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/entity/Counter.java
@@ -1,0 +1,18 @@
+package com.tencent.wxcloudrun.entity;
+
+import lombok.Data;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
+
+@Data
+public class Counter implements Serializable {
+
+  private Integer id;
+
+  private Integer count;
+
+  private LocalDateTime createdAt;
+
+  private LocalDateTime updatedAt;
+}
