+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/mapper/AdsInfoMapper.java
@@ -1,0 +1,7 @@
+package com.tencent.wxcloudrun.mapper;
+
+
+import com.tencent.wxcloudrun.entity.AdsInfoEntity;
+
+public interface AdsInfoMapper extends DefaultMapper<AdsInfoEntity>{
+}
