+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/mapper/CountersMapper.java
@@ -1,0 +1,15 @@
+package com.tencent.wxcloudrun.mapper;
+
+import com.tencent.wxcloudrun.entity.Counter;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+@Mapper
+public interface CountersMapper {
+
+  Counter getCounter(@Param("id") Integer id);
+
+  void upsertCount(Counter counter);
+
+  void clearCount(@Param("id") Integer id);
+}
