+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/mapper/DefaultMapper.java
@@ -1,0 +1,6 @@
+package com.tencent.wxcloudrun.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+public interface DefaultMapper<T> extends BaseMapper<T> {
+}
