+++ wx-cloud-dao/src/main/java/com/tencent/wxcloudrun/repository/AdsInfoRepository.java
@@ -1,0 +1,10 @@
+package com.tencent.wxcloudrun.repository;
+
+import com.tencent.wxcloudrun.entity.AdsInfoEntity;
+import com.tencent.wxcloudrun.mapper.AdsInfoMapper;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public class AdsInfoRepository extends BaseRepository<AdsInfoMapper, AdsInfoEntity> {
+}
