+++ wx-cloud-web/src/main/java/com/tencent/wxcloudrun/WxCloudRunApplication.java
@@ -1,0 +1,14 @@
+package com.tencent.wxcloudrun;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@MapperScan(basePackages = {"com.tencent.wxcloudrun.mapper"})
+public class WxCloudRunApplication {  
+
+  public static void main(String[] args) {
+    SpringApplication.run(WxCloudRunApplication.class, args);
+  }
+}
