+++ wx-cloud-web/src/main/resources/application-dev.yml
@@ -1,0 +1,26 @@
+spring:
+  config:
+    activate:
+      on-profile: dev
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://sh-cynosdbmysql-grp-3upkjhve.sql.tencentcdb.com:25581/wx
+    username: root
+    password: fMQCtjm8
+  mvc:
+    view:
+      suffix: .html
+      prefix: /
+
+server:
+  port: 80
+
+mybatis-plus:
+  configuration:
+    map-underscore-to-camel-case: false
+  global-config:
+    db-config:
+      db-type: MYSQL
+      logic-delete-value: 0
+      logic-not-delete-value: 1
+  mapper-locations: classpath*:/mapper/*.xml
