+++ wx-cloud-web/src/main/resources/application.yml
@@ -1,0 +1,22 @@
+spring:
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: jdbc:mysql://${MYSQL_ADDRESS}/${MYSQL_DATABASE:wx}
+    username: ${MYSQL_USERNAME}
+    password: ${MySQL_PASSWORD}
+  mvc:
+    view:
+      suffix: .html
+      prefix: /
+server:
+  port: 80
+
+mybatis-plus:
+  configuration:
+    map-underscore-to-camel-case: false
+  global-config:
+    db-config:
+      db-type: MYSQL
+      logic-delete-value: 0
+      logic-not-delete-value: 1
+  mapper-locations: classpath*:/mapper/*.xml
