+++ wx-cloud-web/src/main/resources/static/index.html
